+++ android/app/src/main/java/com/daedan/festabook/domain/model/TimeTag.kt
@@ -1,6 +1,6 @@
 package com.daedan.festabook.domain.model
 
 data class TimeTag(
-    val timeTagId: Int,
+    val timeTagId: Long,
     val name: String,
 )
